+++ jest.config.js
@@ -1,10 +1,12 @@
 /** @type {import('jest').Config} */
 module.exports = {
-  preset: "ts-jest",
   testEnvironment: "node",
   roots: ["<rootDir>/tests"],
   moduleNameMapper: {
     "^vscode$": "<rootDir>/__mocks__/vscode.ts",
   },
   clearMocks: true,
+  transform: {
+    "^.+\\.tsx?$": ["ts-jest", { isolatedModules: true }],
+  },
 };